- split SourceMethodSymbol and SourceMethodSymbolBase and inherit SourceViewTypeSymbol.TreeBuilderMethodSymbol from it
- some tweaks and workarounds
