+++ src/Aquila.CodeAnalysis/CodeGen/Symbols/SourceGeneratorSymbol.cs
@@ -1,0 +1,12 @@
+﻿using Aquila.CodeAnalysis.Emit;
+
+namespace Aquila.CodeAnalysis.Symbols
+{
+    partial class SourceGeneratorSymbol
+    {
+        internal void EmitInit(PEModuleBuilder module)
+        {
+            // Don't need any initial emit
+        }
+    }
+};
