+++ src/Aquila.CodeAnalysis/FlowAnalysis/Symbols/SourceMethodSymbolBase.cs
@@ -1,14 +1,8 @@
 ﻿using Aquila.CodeAnalysis.FlowAnalysis;
-using Aquila.CodeAnalysis.Semantics.Graph;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Aquila.CodeAnalysis.Symbols
 {
-    partial class SourceMethodSymbol
+    partial class SourceMethodSymbolBase
     {
         /// <summary>
         /// Method flags lazily collected during code analysis.
