+++ src/Aquila.CodeAnalysis/Semantics/Binder/Binder.cs
@@ -140,20 +140,20 @@ namespace Aquila.CodeAnalysis.Semantics
 
     internal class InMethodBinder : Binder
     {
-        private readonly SourceMethodSymbol _method;
+        private readonly SourceMethodSymbolBase _method;
 
         public InMethodBinder(MethodSymbol method, Binder next) : base(next)
         {
             if(method is null)
                 throw new ArgumentNullException(nameof(method));
             
-            if (method is not SourceMethodSymbol m)
-                throw new ArgumentException($@"The method must be {nameof(SourceMethodSymbol)} type", nameof(method));
+            if (method is not SourceMethodSymbolBase m)
+                throw new ArgumentException($@"The method must be {nameof(SourceMethodSymbolBase)} type", nameof(method));
             
             _method = m;
         }
 
-        public override SourceMethodSymbol Method => _method;
+        public override SourceMethodSymbolBase Method => _method;
 
 
         public override NamespaceOrTypeSymbol Container => _method.ContainingType;
@@ -211,7 +211,7 @@ namespace Aquila.CodeAnalysis.Semantics
 
         public AquilaCompilation Compilation { get; }
 
-        public virtual SourceMethodSymbol Method { get; }
+        public virtual SourceMethodSymbolBase Method { get; }
 
         public virtual NamespaceOrTypeSymbol Container { get; }
